Create WinRDBI Algebra
A recording label has a unique id, a name and a location. Each recording label produces several CD titles. But each CD title may be produced by one and only one recording label. A CD title has a title and a unique code. The year in which the CD was released and the number sold so far must also be recorded.

Each CD title consists of several songs, with a minimum of one song in each CD. Each song has a title and a unique code. The same song could also be present in multiple CD titles. Each song has a unique track number within a particular CD. For each year, the end-of-the-year rating (top 40) of each CD title and song title must be maintained.

A CD title is recorded by a single musical group, which has a name, and a unique code. A musical group may record several CD titles during its life span. The group consists of one or more artists, each of whom has a first name, last name and a unique id. The year in which the artist was born is also recorded. Since any artist may belong to multiple musical groups over time, the database must record the date when an artist joined a particular group and the date when he or she left the group. One or more artists write each song title on a CD. A single artist may contribute towards the writing of multiple song titles within a CD, but it isn’t necessary for an artist to write a song.